renamed lambda functions and added more descriptive error messages.
