+++ amplify/backend/function/deleteUser/Pipfile
