+++ amplify/backend/function/deleteUser/amplify.state
