+++ amplify/backend/function/deleteUser/custom-policies.json
