+++ amplify/backend/function/deleteUser/deleteUser-cloudformation-template.json
@@ -47,12 +47,12 @@
         "FunctionName": {
           "Fn::If": [
             "ShouldNotCreateEnvResources",
-            "neptuneGet",
+            "deleteUser",
             {
               "Fn::Join": [
                 "",
                 [
-                  "neptuneGet",
+                  "deleteUser",
                   "-",
                   {
                     "Ref": "env"
@@ -89,12 +89,12 @@
         "RoleName": {
           "Fn::If": [
             "ShouldNotCreateEnvResources",
-            "yumLambdaRole7677c236",
+            "yumLambdaRoleaccd9a94",
             {
               "Fn::Join": [
                 "",
                 [
-                  "yumLambdaRole7677c236",
+                  "yumLambdaRoleaccd9a94",
                   "-",
                   {
                     "Ref": "env"
