+++ amplify/backend/function/deleteUser/src/event.json
