+++ amplify/backend/function/deleteUser/src/setup.py
