+++ amplify/backend/function/deleteUser/src/src.egg-info/PKG-INFO
