+++ amplify/backend/function/deleteUser/src/src.egg-info/SOURCES.txt
