+++ amplify/backend/function/deleteUser/src/src.egg-info/top_level.txt
