+++ amplify/backend/function/getUser/src/event.json
