+++ amplify/backend/function/getUser/src/index.py
@@ -11,17 +11,21 @@ def lambda_handler(event, context):
     db = client.Client(f"wss://{endpoint}:{port}/gremlin", "g")
 
     username = event["pathParameters"]["userId"]
-    query = f"g.V().hasLabel('user').has('username', '{username}').drop()"
-
+    query = f"g.V().hasLabel('user').has('username', '{username}').valueMap();"
+    
     try:
-        db.submit(query)
-        result = "Query executed successfully"
-        statusCode = 200
+        result = list(db.submit(query))
+        if len(result) == 0:
+            result = f"Server failed to find user : {username}"
+            statusCode = 404
+        else:
+            result = result[0]
+            statusCode = 200
     except:
-        result = "Query failed"
-        statusCode = 400
+        result = "Unknown server error"
+        statusCode = 500
 
-    # Process and return results
+
     return {
         "statusCode": statusCode,
         "body": json.dumps(result)
