+++ amplify/backend/function/getUser/src/src.egg-info/PKG-INFO
