+++ amplify/backend/function/getUser/src/src.egg-info/top_level.txt
