+++ amplify/backend/function/postUser/postUser-cloudformation-template.json
@@ -47,12 +47,12 @@
         "FunctionName": {
           "Fn::If": [
             "ShouldNotCreateEnvResources",
-            "neptunePut",
+            "postUser",
             {
               "Fn::Join": [
                 "",
                 [
-                  "neptunePut",
+                  "postUser",
                   "-",
                   {
                     "Ref": "env"
@@ -89,12 +89,12 @@
         "RoleName": {
           "Fn::If": [
             "ShouldNotCreateEnvResources",
-            "yumLambdaRole34c73d41",
+            "yumLambdaRolec393bd19",
             {
               "Fn::Join": [
                 "",
                 [
-                  "yumLambdaRole34c73d41",
+                  "yumLambdaRolec393bd19",
                   "-",
                   {
                     "Ref": "env"
