+++ amplify/backend/function/postUser/src/event.json
