+++ amplify/backend/function/postUser/src/src.egg-info/PKG-INFO
