+++ amplify/backend/function/postUser/src/src.egg-info/top_level.txt
