+++ amplify/backend/function/putUser/src/index.py
@@ -1,0 +1,35 @@
+import json
+import os
+import gremlin_python
+from gremlin_python.driver import client, serializer
+
+def lambda_handler(event, context):
+    endpoint = os.environ['NEPTUNE_ENDPOINT']
+    port = os.environ['NEPTUNE_PORT']
+
+    # Connect to Neptune
+    db = client.Client(f"wss://{endpoint}:{port}/gremlin", "g")
+
+    userData = json.loads(event["body"])["userInfo"]
+    username = event["pathParameters"]["userId"]
+    attribute = userData["attribute"]
+    newValue = userData["newValue"]
+    # Get vertex ID
+    query = f"g.V().has('username', '{username}').id();"
+    vertexID = db.submit(query).all().result()
+    # Add new value
+    query = f"g.V({vertexID}).property(single, '{attribute}', '{newValue}');"
+
+    try:
+        db.submit(query)
+        result = f"User attribute : {attribute} successfully changed to : {newValue}"
+        statusCode = 200
+    except:
+        result = f"Server failed at modifying attribute : {attribute} for user : {username}"
+        statusCode = 500
+
+    # Process and return results
+    return {
+        "statusCode": statusCode,
+        "body": json.dumps(result)
+    }
